Fix typo, rename tag reason, correctly handle uninitialized key value
Rename {Rcv,Snd}bufErrors tag to the same value: nomem.

Also, we know RcvbufErrors key exists, so just refer to it using the normal dictionary notation, no need to provide a default value. However, always provide a default if we're not 100% sure a key exists, in this case for InErrors (InErrors are also used for Icmp, which is why we rely on RcvbufErrors as a good indicator that we're handling Udp related stats).