added option for config-refresh-interval
Config files are normally reloaded if they have
changed their modified time.  However, it's
possible that tag or option values are dynamic
and do not depend on the config file changing.

In this case we want to be able to set an interval
at which the files will be updated anyway.

This is controlled with --config-refresh-interval